Tile page getting there
The photo tile page is getting there, but with a few lingering issues:

- Colors are awful and I can't figure out how to set them using
variables for ease with LESS unless it needs to be downloaded
separately from bootstrap

- To get that "jagged grid" look like lulu (which i think is critical
to the infinite scrolling design) you have to create new rows within
columns.

Problem is that when you start to put content in a new row vs a new
column changes with screen size as there are a different number of
columns for xs, sm, md and lg/xl devices.

This affects the design as a new column that is stacked horizontally
below an existing column will not allow for that "jagged grid" layout.
Rather the top of that column will begin after the lowest point in the
columns above.

One solution I have thought of but would have to be implemented with
other coding (potentially on the backend later) is:

(1) using a media query to actually make columns disappear with
different screen sizes
(2) take the thumbnails which would have been included and add them to
rows in visible columns instead